requirements.txt
# Core dependencies
pymongo>=4.6.1
requests>=2.31.0
beautifulsoup4>=4.12.2
fake-useragent>=1.4.0
python-whois>=0.8.0
dnspython>=2.5.0
urllib3<2.0.0
certifi>=2023.11.17

# HTML Processing
lxml>=4.9.3
html5lib>=1.1

# Date/Time Handling
python-dateutil>=2.8.2
pytz>=2023.3

# Network/SSL
cryptography>=41.0.7
pyOpenSSL>=23.3.0

# Error Handling & Logging
python-json-logger>=2.0.7

# Type Hints
typing-extensions>=4.9.0

# Concurrent Processing
futures>=3.0.5

# Data Processing
pandas>=2.1.4

# Optional but recommended
aiohttp>=3.9.1
tld>=0.13.0